working wand added!
