+++ index.html
@@ -12,7 +12,7 @@
     <link
         href="https://fonts.googleapis.com/css2?family=Bitter:ital,wght@0,400;0,500;0,800;1,400;1,500;1,600;1,800&display=swap"
         rel="stylesheet">
-    <script src="https://kit.fontawesome.com/34d22bbab1.js" crossorigin="anonymous"></script>
+<!--    <script src="https://kit.fontawesome.com/34d22bbab1.js" crossorigin="anonymous"></script>-->
 </head>
 
 <body>
@@ -31,16 +31,16 @@
         </main>
         <footer class="footer">
             <div class="game-timer">
-                <img src="./static/assets/img/wand.png" alt="The shaft of the wand">
-                <div class="game-timer__wand">
-                    <div class="inner"></div>
-                </div>
+<!--                <img src="./static/assets/img/wand.png" alt="The shaft of the wand">-->
+<!--                <div class="game-timer__wand">-->
+<!--                    <div class="inner"></div>-->
+<!--                </div>-->
             </div>
-            <button class="start">START</button>
+
         </footer>
     </div>
     <script type="module" src="./src/index.js"></script>
-    <script src="./src/wand.js"></script>
+<!--    <script src="src/app/view/components/wand.js"></script>-->
 </body>
 
 </html>
