+++ src/app/data/config.js
@@ -1,1 +1,1 @@
-export const GAME_TIME = 0.5;
+export const GAME_TIME = 10 / 60; // ?seconds / 60seconds
